requirements.txt
streamlit
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain-huggingface
langchain-community
huggingface-hub
langchain_google_genai 
google-generativeai
